The rectangles a tree painted are asked for in one place
`no_caret.rs` had a ten-line walker over `RenderNode` collecting every
`RoundedRect`, because that is what a caret is: a rect with no radius.
The tests for editing want the same walker, to ask where the caret
ended up after the view scrolled — and a second byte-identical copy is
the drift `tests/common/mod.rs` was written to stop.

`Harness::painted_rects` is that walker, beside the `paint` it already
owns. `no_caret.rs` loses its copy, its hand-rolled tree and its
`Constraints` import along with it.
